backend/upload_api/requirements.txt
google-cloud-aiplatform
google-cloud-firestore
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3